[BUGFIX] Only remove and recreate directories on low level cache flush
When removing cache directories, we must ignore files,
as those are removed when removing the directory anyway.

+++ Classes/Console/Service/CacheLowLevelCleaner.php
@@ -29,6 +29,9 @@ class CacheLowLevelCleaner
     {
         $cacheDirPattern = Environment::getVarPath() . '/cache/*/*';
         foreach (glob($cacheDirPattern) as $path) {
+            if (!is_dir($path)) {
+                continue;
+            }
             GeneralUtility::rmdir($path, true);
             GeneralUtility::mkdir($path);
         }
